+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/Internals/TransmitFromStorageHandler.cs
@@ -54,7 +54,7 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals
                             _transmissionStateManager.ResetConsecutiveErrors();
                             _transmissionStateManager.CloseTransmission();
 
-                            AzureMonitorExporterEventSource.Log.TransmitFromStorageSuccess();
+                            AzureMonitorExporterEventSource.Log.TransmitFromStorageSuccess(_connectionVars.InstrumentationKey);
 
                             // In case if the delete fails, there is a possibility
                             // that the current batch will be transmitted more than once resulting in duplicates.
@@ -70,7 +70,7 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals
                     }
                     catch (Exception ex)
                     {
-                        AzureMonitorExporterEventSource.Log.FailedToTransmitFromStorage(ex);
+                        AzureMonitorExporterEventSource.Log.FailedToTransmitFromStorage(_connectionVars.InstrumentationKey, ex);
                     }
                 }
                 else
